Cosmetic Changes
1. In this commit all constant aliasings / redeclarations of the
   variables of Internal Vars are declared using 'auto'
2. Changed declaration of STDPMaxTimeWin to use an 'int' argument
   as opposed to size_t.
3. Changed some doubles to floats to avoid type compatibility
   warnings.

+++ TimeDelNetSimMEX_Lib/Source/MexFunctionInterface.cpp
@@ -209,7 +209,7 @@ void takeInputFromMatlabStruct(mxArray* MatlabInputStruct, InputArgs &InputArgLi
 	// Initializing IExtGenState
 	genmxArrayPtr = mxGetField(MatlabInputStruct, 0, "IExtGenState");
 	if (genmxArrayPtr != NULL && !mxIsEmpty(genmxArrayPtr)){
-		int NElems = mxGetNumberOfElements(genmxArrayPtr);
+		size_t NElems = mxGetNumberOfElements(genmxArrayPtr);
 		InputArgList.IExtGenState = MexVector<uint32_t>(NElems);
 		genUIntPtr[0] = reinterpret_cast<uint32_t *>(mxGetData(genmxArrayPtr));
 		InputArgList.IExtGenState.copyArray(0, genUIntPtr[0], NElems);
